Update to cmake 3.21 (#63)
* Some updates

* Fix warnings from distconv utils

* Fix some warnings

* Fix some CUDA-related variables

* Fix a bunch of warnings and/or errors

* Correct some format strings

* Cleanup the CMake export file

* Fix cuDNN stuff

* Address shared library issue from review

* Fix NVSHMEM module

* Update softmax.cu

+++ legacy/include/distconv/tensor/tensor_mpi.hpp
@@ -185,7 +185,7 @@ class TensorImpl<Tensor<DataType, LocaleMPI, Allocator>> {
     auto &offsets = m_offset_all.back();
     index_t cur_offset = 0;
     for (int i = 0; i < num_procs; i += num_procs_per_split) {
-      for (int j = 0; j < num_procs_per_split; ++j) {
+      for (size_t j = 0; j < num_procs_per_split; ++j) {
         offsets[i+j] = cur_offset;
       }
       cur_offset += local_dim;
